# agricultural/api-tests/apifox-openapi-import.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "助农电商平台 API",
    "description": "用户端 / 商家端 / 管理员端 RESTful 接口 - 助农电商平台",
    "contact": {
      "name": "助农平台技术支持",
      "email": "support@agricultural.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/user/products/search": {
      "get": {
        "tags": ["用户端-商品"],
        "summary": "商品搜索",
        "operationId": "searchProducts",
        "parameters": [
          {"name": "keyword", "in": "query", "schema": {"type": "string"}},
          {"name": "sortBy", "in": "query", "schema": {"type": "string", "enum": ["sales", "price", "score"]}},
          {"name": "categoryId", "in": "query", "schema": {"type": "integer"}},
          {"name": "originPlace", "in": "query", "schema": {"type": "string"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultProductSimpleVO"}}}}
        }
      }
    },
    "/user/products/{id}": {
      "get": {
        "tags": ["用户端-商品"],
        "summary": "商品详情",
        "operationId": "productDetail",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultProductDetailVO"}}}}
        }
      }
    },
    "/user/products/hot": {
      "get": {
        "tags": ["用户端-商品"],
        "summary": "热销榜单",
        "operationId": "hotProducts",
        "parameters": [
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ProductSimpleVO"}}}}}
        }
      }
    },
...
    "/user/merchants/{id}": {
      "get": {
        "tags": ["用户端-商家"],
        "summary": "商家主页",
        "operationId": "merchantDetail",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultShopInfo"}}}}
        }
      }
    },
    "/user/merchants/{id}/products": {
      "get": {
        "tags": ["用户端-商家"],
        "summary": "商家商品列表",
        "operationId": "merchantProducts",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultProductSimpleVO"}}}}
        }
      }
    },
...
    "/user/comments": {
      "get": {
        "tags": ["用户端-评价"],
        "summary": "评价列表",
        "operationId": "commentList",
        "parameters": [
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultCommentVO"}}}}
        }
      },
      "post": {
        "tags": ["用户端-评价"],
        "summary": "提交评价",
        "operationId": "submitComment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderNo": {"type": "string"},
                  "productId": {"type": "integer"},
                  "score": {"type": "integer", "description": "1-5分"},
                  "content": {"type": "string"},
                  "imgUrls": {"type": "string"}
                },
                "required": ["orderNo", "productId", "score"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"type": "object"}
        }
      },
...
      "ProductSimpleVO": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "productName": {"type": "string"},
          "price": {"type": "number"},
          "productImg": {"type": "string"},
          "salesVolume": {"type": "integer"},
          "stock": {"type": "integer"}
        }
      },
      "PageResultProductSimpleVO": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/ProductSimpleVO"}}
        }
      },
...
      "ResultProductDetailVO": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/ProductDetailVO"}
        }
      },
      "ProductDetailVO": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "productName": {"type": "string"},
          "price": {"type": "number"},
          "stock": {"type": "integer"},
          "productImg": {"type": "string"},
          "productDesc": {"type": "string"},
          "originPlace": {"type": "string"},
          "salesVolume": {"type": "integer"},
          "score": {"type": "number"},
          "merchantId": {"type": "integer", "format": "int64"},
          "merchantName": {"type": "string"},
          "traceInfo": {"$ref": "#/components/schemas/ProductTraceVO"}
        }
      },
      "ProductTraceVO": {
        "type": "object",
        "properties": {
          "plantingCycle": {"type": "string"},
          "originPlaceDetail": {"type": "string"},
          "fertilizerType": {"type": "string"},
          "storageMethod": {"type": "string"},
          "transportMethod": {"type": "string"},
          "qrCodeUrl": {"type": "string"}
        }
      },
...
      "ResultShopInfo": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/ShopInfo"}
        }
      },
      "ShopInfo": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "merchantId": {"type": "integer", "format": "int64"},
          "shopName": {"type": "string"},
          "shopIntro": {"type": "string"},
          "qualificationImg": {"type": "string"}
        }
      },
...
      "CommentVO": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "orderNo": {"type": "string"},
          "productId": {"type": "integer", "format": "int64"},
          "productName": {"type": "string"},
          "score": {"type": "integer"},
          "content": {"type": "string"},
          "imgUrls": {"type": "string"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
      "PageResultCommentVO": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/CommentVO"}}
        }
      },
...
    }
  }
}